+++ module/effects/effects-config.js
@@ -4,8 +4,8 @@ export default class ActiveEffectConfig4e extends ActiveEffectConfig {
 	static get defaultOptions() {
 		return foundry.utils.mergeObject(super.defaultOptions, {
 			classes: ["sheet", "active-effect-sheet"],
-			template: "templates/sheets/active-effect-config.html",
-			width: 260,
+			template: "systems/dnd4e/templates/sheets/active-effect-config.html",
+			width: 560,
 			height: "auto",
 			tabs: [{navSelector: ".tabs", contentSelector: "form", initial: "details"}]
 		});