+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - item sheet css fix
 - added some more sorting options for powers
 - powers will also have a tertiary alphabetical sorting applied to them.
+- active effects provided by items now have an option wherre the effect is only turned on while the item is equipted
 
 ## Version 0.2.73
 - fix damage type calculation