+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - active effects from powers can be applyed to target tokens
 - duration options have been added to active effects
 - when half levels are turned off, it will no longer show the value of mod + 1/2 level next to the ability mods.
+- The value for Area of Effect sizes within powers/consumables is now a string text, and can be composed of basic roll formulas.
 
 ## Version 0.2.73
 - fix damage type calculation